+++ frontend/src/components/Recruit/style.jsx
@@ -34,6 +34,38 @@ export const StartEndTimeBox = styled.div`
 `;
 
 /* TimeSelect */
+export const ModalBackground = styled.div`
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100vw;
+  height: 100vh;
+  background: rgba(0, 0, 0, 0.3);
+  z-index: 99;
+`;
+export const ModalBox = styled.div`
+  transform: translate(-50%, -50%);
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  background-color: white;
+  width: 50rem;
+  height: 30rem;
+`;
+export const ModalHeader = styled.div`
+  display: flex;
+  align-items: center;
+  height: 2rem;
+  border-bottom: 0.5px solid rgba(0, 0, 0, 0.3);
+`;
+export const CloseModalBtn = styled.button`
+  margin-left: auto;
+  margin-right: 0.5rem;
+`;
+export const ModalContent = styled.div`
+  display: flex;
+`;
+export const CalendarView = styled.div``;
 export const GroupBox = styled.div``;
 export const GroupBtn = styled.button`
   width: 2rem;
